修复保存insert时出错bug

+++ soafw-common-dao/src/main/java/com/kjt/service/common/dao/ibatis/IBigIMapper.java
@@ -15,6 +15,6 @@ public interface IBigIMapper<T extends IModel> extends IMapper<T> {
 
   public T queryById(BigInteger id);
 
-  public BigInteger insert(IModel params);
+  public Long insert(IModel params);
 
 }
