+++ soafw-common-dao/src/main/java/com/kjt/service/common/dao/ibatis/IIMapper.java
@@ -13,6 +13,6 @@ public interface IIMapper<T extends IModel> extends IMapper<T> {
 
   public T queryById(Integer id);
 
-  public Integer insert(T model);
+  public Long insert(T model);
 
 }
